refactor: postComment - reader, writer
