+++ src/module/post-comment/post-comment.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { PostCommentController } from './post-comment.controller';
 import { PostCommentService } from './post-comment.service';
-import { PostCommentRepository } from './repository/post-comment.repository';
+import { PostCommentReader } from './repository/post-comment.reader';
+import { PostCommentWriter } from './repository/post-comment.writer';
 import { AwsModule } from '../aws/aws.module';
 
 @Module({
   imports: [AwsModule],
   controllers: [PostCommentController],
-  providers: [PostCommentService, PostCommentRepository],
+  providers: [PostCommentService, PostCommentReader, PostCommentWriter],
 })
 export class PostCommentModule {}
